requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.1
django-cors-headers==3.0.2
django-cors-middleware==1.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.4
djangorestframework==3.9.4
et-xmlfile==1.0.1
future==0.17.1
httpie==1.0.2
httplib2==0.9.2
idna==2.8
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
Pygments==2.4.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
